add custom font colors for inline laels
